Move back from the adapter zone per conversation with Chenyu.
